+++ src/sync_type.rs
@@ -6,7 +6,7 @@ use std::ptr::NonNull;
 
 use crate::counter::Counter;
 use crate::handle::RefBox;
-use crate::pointer::Pointer;
+use crate::pointer::{NullablePointer, Pointer};
 use crate::pool::PoolInner;
 use crate::stack::Stack;
 
@@ -15,14 +15,14 @@ use crate::pointer::NonNullAtomicPtr;
 #[cfg(feature = "sync")]
 use crossbeam_queue::ArrayQueue;
 #[cfg(feature = "sync")]
-use std::sync::atomic::AtomicUsize;
+use std::sync::atomic::{AtomicPtr, AtomicUsize};
 
 #[doc(hidden)]
 pub trait PoolSyncType<A>: Sized {
     type Counter: Counter;
     type Stack: Stack<Self::ElementPointer>;
     type ElementPointer: Pointer<RefBox<A, Self>>;
-    type PoolPointer: Pointer<PoolInner<A, Self>>;
+    type PoolPointer: NullablePointer<PoolInner<A, Self>>;
 }
 
 /// Marker type for thread safe pools.
@@ -39,12 +39,12 @@ impl<A> PoolSyncType<A> for PoolSync {
     type Counter = AtomicUsize;
     type Stack = ArrayQueue<Self::ElementPointer>;
     type ElementPointer = NonNullAtomicPtr<RefBox<A, Self>>;
-    type PoolPointer = NonNullAtomicPtr<PoolInner<A, Self>>;
+    type PoolPointer = AtomicPtr<PoolInner<A, Self>>;
 }
 
 impl<A> PoolSyncType<A> for PoolUnsync {
     type Counter = usize;
     type Stack = Vec<Self::ElementPointer>;
     type ElementPointer = NonNull<RefBox<A, Self>>;
-    type PoolPointer = NonNull<PoolInner<A, Self>>;
+    type PoolPointer = *mut PoolInner<A, Self>;
 }